pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "setuptools_scm[toml]>=6.0"]
build-backend = "setuptools.build_meta"


[project]
name = "streamcondor"
dynamic = ["version"]
description = "A system tray application for monitoring livestreams status"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    { name = "Tarzasai" },
]
keywords = ["stream", "tray", "monitor", "pyqt6"]
classifiers = [
	"Development Status :: 4 - Beta",
	"Intended Audience :: End Users/Desktop",
	"Programming Language :: Python :: 3",
	"Programming Language :: Python :: 3.12",
	"Operating System :: OS Independent",
]

dependencies = [
	"PyQt6>=6.10.0",
	"streamlink>=8.0.0",
	"requests>=2.32.5",
	"beautifulsoup4>=4.14.2",
	"Pillow>=12.0.0",
	"pyperclip>=1.11.0",
	"pydantic>=2.12.5",
]

[project.scripts]
streamcondor = "streamcondor.main:main"

[tool.setuptools_scm]
version_scheme = "post-release"
local_scheme = "no-local-version"
write_to = "src/streamcondor/_version.py"
